+++ src/test/java/bumblebee/xchangepass/domain/user/service/UserLoginScenarioTest.java
@@ -6,13 +6,13 @@ import bumblebee.xchangepass.domain.user.entity.Role;
 import bumblebee.xchangepass.domain.user.entity.Sex;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
 import bumblebee.xchangepass.global.security.jwt.JwtProvider;
-import bumblebee.xchangepass.global.security.v1.login.LoginService;
-import bumblebee.xchangepass.global.security.v1.login.UserRegisterService;
-import bumblebee.xchangepass.global.security.v1.login.dto.LoginRequest;
-import bumblebee.xchangepass.global.security.v1.login.dto.LoginResponse;
-import bumblebee.xchangepass.global.security.v1.refresh.RefreshToken;
-import bumblebee.xchangepass.global.security.v1.refresh.RefreshTokenService;
-import bumblebee.xchangepass.global.security.v1.refresh.dto.RefreshTokenResponse;
+import bumblebee.xchangepass.global.security.login.LoginService;
+import bumblebee.xchangepass.global.security.login.UserRegisterService;
+import bumblebee.xchangepass.global.security.login.dto.LoginRequest;
+import bumblebee.xchangepass.global.security.login.dto.LoginResponse;
+import bumblebee.xchangepass.global.security.refresh.RefreshToken;
+import bumblebee.xchangepass.global.security.refresh.RefreshTokenService;
+import bumblebee.xchangepass.global.security.refresh.dto.RefreshTokenResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
