+++ src/test/java/bumblebee/xchangepass/domain/user/service/UserServiceTest.java
@@ -9,7 +9,7 @@ import bumblebee.xchangepass.domain.user.repository.UserRepository;
 import bumblebee.xchangepass.global.error.ErrorCode;
 import bumblebee.xchangepass.global.exception.CommonException;
 import bumblebee.xchangepass.global.scheduler.UserCleanupScheduler;
-import bumblebee.xchangepass.global.security.v1.login.UserRegisterService;
+import bumblebee.xchangepass.global.security.login.UserRegisterService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
